+++ custom_components/maintenance_supporter/frontend/panel-chunks/object-dialog-LYWSTV3W.js
@@ -1,0 +1,2 @@
+/*! maintenance_supporter frontend 2.88.0 */
+import{a}from"/maintenance_supporter_panelfiles/panel-chunks/chunk-7H4RTG6L.js";import"/maintenance_supporter_panelfiles/panel-chunks/chunk-X7WY7CJL.js";import"/maintenance_supporter_panelfiles/panel-chunks/chunk-I72OGKOI.js";import"/maintenance_supporter_panelfiles/panel-chunks/chunk-IWWLTDMV.js";import"/maintenance_supporter_panelfiles/panel-chunks/chunk-WH2PUYFW.js";export{a as MaintenanceObjectDialog};